+++ src/django_shared_property/expressions.py
@@ -1,16 +1,16 @@
-from django.db.models.expressions import Expression, F
+from django.db.models.expressions import Expression
 from django.db.models.sql.query import Query
 
 
 class ExpressionCol(Expression):
     contains_aggregate = False
 
-    def __init__(self, expression, model, alias=None, output_field=None):
+    def __init__(self, expression, model, alias=None, output_field=None, target=None):
         self.expression = expression
-        if output_field:
-            self.output_field = output_field
+        self.output_field = output_field or expression.resolve_expression(Query(model)).output_field
         self.alias = alias
         self.model = model
+        self.target = target
 
     def get_lookup(self, name):
         return self.output_field.get_lookup(name)
@@ -23,7 +23,4 @@ class ExpressionCol(Expression):
         return resolved.as_sql(compiler, connection)
 
     def get_db_converters(self, connection):
-        return self.output_field.get_db_converters(connection) + self.expression.get_db_converters(connection)
-
-    def resolve_expression(self, *args, **kwargs):
-        return self.expression.resolve_expression(Query(self.model))
+        return self.output_field.get_db_converters(connection)  # + self.expression.get_db_converters(connection)