+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+0.2.0 (2021-07-22)
+------------------
+
+* Support Enum return values.
+* Simplify decorator code.
+
+
 0.1.0 (2020-09-15)
 ------------------